+++ azure_monitor/src/azure_monitor/export/__init__.py
@@ -4,11 +4,16 @@ import json
 import logging
 import typing
 from enum import Enum
+from urllib.parse import urlparse
 
 import requests
 from opentelemetry.sdk.metrics.export import MetricsExportResult
 from opentelemetry.sdk.trace.export import SpanExportResult
+from opentelemetry.sdk.util import ns_to_iso_str
+from opentelemetry.trace import Span, SpanKind
+from opentelemetry.trace.status import StatusCanonicalCode
 
+from azure_monitor import protocol, utils
 from azure_monitor.options import ExporterOptions
 from azure_monitor.protocol import Envelope
 from azure_monitor.storage import LocalFileStorage
@@ -90,15 +95,16 @@ class BaseExporter:
             # give a few more seconds for blob lease operation
             # to reduce the chance of race (for perf consideration)
             if blob.lease(self.options.timeout + 5):
-                envelopes = blob.get()  # TODO: handle error
+                envelopes = blob.get()
                 result = self._transmit(envelopes)
                 if result == ExportResult.FAILED_RETRYABLE:
                     blob.lease(1)
                 else:
-                    blob.delete(silent=True)
+                    blob.delete()
 
     # pylint: disable=too-many-branches
     # pylint: disable=too-many-nested-blocks
+    # pylint: disable=too-many-return-statements
     def _transmit(self, envelopes: typing.List[Envelope]) -> ExportResult:
         """
         Transmit the data envelopes to the ingestion service.
@@ -116,9 +122,18 @@ class BaseExporter:
                         "Content-Type": "application/json; charset=utf-8",
                     },
                     timeout=self.options.timeout,
+                    proxies=self.options.proxies,
                 )
+            except requests.Timeout:
+                logger.warning(
+                    "Request time out. Ingestion may be backed up. Retrying."
+                )
+                return ExportResult.FAILED_RETRYABLE
             except Exception as ex:
-                logger.warning("Transient client side error: %s.", ex)
+                logger.warning(
+                    "Retrying due to transient client side error %s.", ex
+                )
+                # client side error (retryable)
                 return ExportResult.FAILED_RETRYABLE
 
             text = "N/A"